Removed first approach of Coroutines tests. Added new approach of tests.
Added "android.arch.core:core-testing:1.1.1" dependency for the InstantTaskExecutorRule.

Added a database test module with inMemoryDatabaseBuilder.

Added baseUnitTest class.

Added OneTimeObserver with the extension observeOnce to easily and efficiently test the LiveData.

Modify searchArtistMethod on ArtistSearchViewModel, now return a Job in order to use .join()(wait for completion) in the tests.